DocumentWriterModal 리팩토링
