+++ app/approvals/components/DocumentWriterModal/components/CollapsibleSection.tsx
@@ -1,0 +1,35 @@
+"use client"
+
+import { useState } from "react"
+import { ChevronDown, ChevronUp } from "lucide-react"
+import { typography } from "@/lib/design-tokens"
+import { CollapsibleSectionProps } from "../types"
+
+export function CollapsibleSection({
+  title,
+  children,
+  defaultOpen = false
+}: CollapsibleSectionProps) {
+  const [isOpen, setIsOpen] = useState(defaultOpen)
+
+  return (
+    <div>
+      <button
+        onClick={() => setIsOpen(!isOpen)}
+        className="w-full flex items-center justify-between p-4 text-left hover:bg-gray-50 transition-colors"
+      >
+        <h3 className={`${typography.h4} text-gray-800`}>{title}</h3>
+        {isOpen ? (
+          <ChevronUp className="w-5 h-5 text-gray-500" />
+        ) : (
+          <ChevronDown className="w-5 h-5 text-gray-500" />
+        )}
+      </button>
+      {isOpen && (
+        <div className="px-4 pb-4">
+          {children}
+        </div>
+      )}
+    </div>
+  )
+}
